Implement compilation to file
Also added options:
     -s to save temporary files
     -d to dump assembler to stdout post-generation

+++ src/rattle.c
@@ -53,52 +53,63 @@ void evaluate(const char *);
 void compile(const char *, const char *);
 void compile_expression (const char *);
 
+// TODO find correct posix value
+#define FILE_PATH_MAX 1024
+
+// Option globals
+static bool dump_p = false;
+static bool save_temps_p = false;
+
 int
-main(int argc, char *argv[]) {
+main (int argc, char *argv[]) {
 
   bool compile_p = false;
   bool evaluate_p = false;
-  char *input = NULL;
-  char *output = NULL;
+  char input[FILE_PATH_MAX];
+  char output[FILE_PATH_MAX];
 
   int opt;
-  while ((opt = getopt(argc, argv, "hec:o:")) != -1)
+  while ((opt = getopt(argc, argv, "hdsec:o:")) != -1)
     {
       switch (opt)
         {
         case 'h':
-          help(argv[0]);
+          help (argv[0]);
+          break;
+        case 'd':
+          dump_p = true;
+          break;
+        case 's':
+          save_temps_p = true;
           break;
         case 'c':
           compile_p = true;
-          assert (false);
-          //input = strdup(optarg);
+          strcpy (input, optarg);
           break;
         case 'o':
-          assert (false);
-          //output = strdup(optarg);
+          strcpy (output, optarg);
           break;
         case 'e':
           evaluate_p = true;
           break;
         case ':':
           fprintf (stderr, "flag missing operand\n");
-          usage(argv[0]);
+          usage (argv[0]);
           break;
         case '?':
           fprintf (stderr, "unrecognized option: '-%c'\n", optopt);
-          usage(argv[0]);
+          usage (argv[0]);
           break;
         default:
-          usage(argv[0]);
+          usage (argv[0]);
           break;
         }
     }
 
   if (evaluate_p && compile_p)
     {
-      fprintf(stderr, "cannot specify both -e and -c\n");
-      usage(argv[0]);
+      fprintf (stderr, "cannot specify both -e and -c\n");
+      usage (argv[0]);
     }
 
   if (evaluate_p)
@@ -111,11 +122,11 @@ main(int argc, char *argv[]) {
         }
 
       const char *cmd = argv[optind];
-      evaluate(cmd);
+      evaluate (cmd);
     }
 
   if (compile_p)
-      compile(input, output);
+    compile (input, output);
 
   return 0;
 }
@@ -702,14 +713,72 @@ emit_asm_prim_nullp (FILE *f, schprim_t *p __attribute__((unused)))
   fprintf (f, "    orq    $%" PRIu64 ", %%rax\n", BOOL_TAG);
 }
 
+
 ///////////////////////////////////////////////////////////////////////
 //
-// Section Compilation and Evalution
+// Section Error Management
+//
+// Error management - this should be the only section causing
+// the program to exit
+//
+///////////////////////////////////////////////////////////////////////
+
+void
+err_oom (void)
+{
+  fprintf (stderr, "out of memory\n");
+  exit (EXIT_FAILURE);
+}
+
+void
+err_parse (const char *s)
+{
+  fprintf (stderr, "error: cannot parse `%s'\n", s);
+  exit (EXIT_FAILURE);
+}
+
+///////////////////////////////////////////////////////////////////////
+//
+// Section Memory Management
+//
+// Memory management - this should be the only section causing the
+// program to allocate memory
+//
+///////////////////////////////////////////////////////////////////////
+
+// alloc: allocates n bytes of memory and returns a pointer to it
+void *
+alloc (size_t n)
+{
+  void *mem = malloc (n);
+  if (!mem)
+    err_oom ();
+
+  return mem;
+}
+
+// grow: grows region pointed to by ptr to n bytes
+void *
+grow (void *ptr, size_t n)
+{
+  void *mem = realloc (ptr, n);
+  if (!mem)
+    err_oom ();
+
+  return mem;
+}
+
+///////////////////////////////////////////////////////////////////////
+//
+// Section Compilation and Evaluation
 //
 // Top-level compilation and evaluation functions.
 //
 //
 ///////////////////////////////////////////////////////////////////////
+
+// Prototypes
+const char *find_system_tmpdir (void);
 
 // Evaluation
 void
@@ -718,16 +787,131 @@ evaluate(const char *cmd)
   compile_expression(cmd);
 }
 
-void
-compile(const char *input, const char *output)
+char *
+output_asm (schptr_t sptr)
 {
-  (void) input;
-  (void) output;
-  assert(false);
+  const char *tmpdir = find_system_tmpdir ();
+  char itemplate[1024];
+  strcpy (itemplate, tmpdir);
+  strcat (itemplate, "/rattleXXXXXX.s");
+
+  int ifd = mkstemps (itemplate, 2);
+  if (ifd == -1)
+    {
+      fprintf (stderr, "error creating temporary files for compilation\n");
+      exit (EXIT_FAILURE);
+    }
+
+    FILE *i = fdopen (ifd, "w");
+  if (!i)
+    {
+      fprintf (stderr, "cannot open file descriptor for `%s'\n", itemplate);
+      exit (EXIT_FAILURE);
+    }
+
+  // write asm file
+  emit_asm_prologue (i, "scheme_entry");
+  emit_asm_expr (i, sptr);
+  emit_asm_epilogue (i);
+
+  // close file
+  fclose (i);
+
+  char *ipath = strdup (itemplate);
+  if (!ipath)
+    err_oom ();
+
+  return ipath;
+}
+
+char *
+read_file_to_mem (const char *path)
+{
+  FILE *f = fopen (path, "r");
+  if (!f)
+    {
+      fprintf (stderr, "cannot open `%s' for reading", path);
+      exit (EXIT_FAILURE);
+    }
+
+  // read file contents to memory
+  const size_t blocksize = 1024;
+  size_t ssize = blocksize;
+  char *s = (char *) alloc (sizeof (*s) * ssize);
+  size_t bytes_read = 0;
+  int c;
+  while ((c = fgetc (f)) != EOF)
+    {
+      s[bytes_read++] = c;
+
+      // maybe grow s?
+      if (bytes_read == ssize)
+        {
+          ssize += blocksize;
+          s = grow (s, ssize);
+        }
+    }
+
+  // close file
+  fclose (f);
+  return s;
+}
+
+void
+dump_asm_if_needed (const char *path)
+{
+  if (dump_p)
+    {
+      char *asmdump = read_file_to_mem (path);
+      printf ("Assembly dump:\n");
+      for (size_t i = 0; i < strlen (asmdump); i++)
+        fputc (asmdump[i], stdout);
+      printf ("End of Assembly dump\n");
+      free (asmdump);
+    }
+}
+
+void
+compile (const char *input, const char *output)
+{
+  char *s = read_file_to_mem (input);
+  schptr_t sptr = 0;
+  const char *cs = s;
+  if (!parse_expr (&cs, &sptr))
+    err_parse (cs);
+
+  // free parsed string
+  free (s);
+
+  char *asmtmp = output_asm (sptr);
+  dump_asm_if_needed (asmtmp);
+
+  // Now compile file and link with runtime
+  {
+    int child = fork ();
+    if (child == 0)
+      {
+        // inside child
+        execl (CC, CC, "-o", output, asmtmp, "runtime.o", (char *) NULL);
+
+        // unreachable
+        assert (false);
+      }
+
+    // wait for child to complete compilation
+    waitpid (child, NULL, 0);
+  }
+
+  // free path to tmp file
+  if (save_temps_p)
+    printf ("Temporary asm source kept at `%s'\n", asmtmp);
+  else
+    unlink (asmtmp);
+  free (asmtmp);
 }
 
 const char *
-find_system_tmpdir ()
+find_system_tmpdir (void)
 {
   static const char *default_tmpdir = "/tmp/";
 
@@ -752,45 +936,23 @@ void
 compile_expression (const char *e)
 {
   schptr_t sptr = 0;
-  bool parsed = parse_expr (&e, &sptr);
-
-  if (!parsed)
-    {
-      fprintf (stderr, "error: cannot parse `%s'\n", e);
-      exit (EXIT_FAILURE);
-    }
+  if (!parse_expr (&e, &sptr))
+    err_parse (e);
 
   const char *tmpdir = find_system_tmpdir ();
-  char itemplate[1024];
   char otemplate[1024];
-  strcpy (itemplate, tmpdir);
-  strcat (itemplate, "/rattleXXXXXX.s");
   strcpy (otemplate, tmpdir);
   strcat (otemplate, "/librattleXXXXXX.so");
 
-  int ifd = mkstemps (itemplate, 2);
   int ofd = mkstemps (otemplate, 3);
-
-  if (ifd == -1 || ofd == -1)
+  if (ofd == -1)
     {
       fprintf (stderr, "error creating temporary files for compilation\n");
       exit (EXIT_FAILURE);
     }
 
-  FILE *i = fdopen (ifd, "w");
-  if (!i)
-    {
-      fprintf (stderr, "cannot open file descriptor for `%s'\n", itemplate);
-      exit (EXIT_FAILURE);
-    }
-
-  // write asm file
-  emit_asm_prologue (i, "scheme_entry");
-  emit_asm_expr (i, sptr);
-  emit_asm_epilogue (i);
-
-  // close file
-  fclose (i);
+  char *asmtmp = output_asm (sptr);
+  dump_asm_if_needed (asmtmp);
 
   // close ofd so gcc can write to it
   close (ofd);
@@ -801,7 +963,7 @@ compile_expression (const char *e)
     if (child == 0)
       {
         // inside child
-        execl (CC, CC, "-shared", "-fPIC", "-o", otemplate, itemplate, "runtime.o", (char *) NULL);
+        execl (CC, CC, "-shared", "-fPIC", "-o", otemplate, asmtmp, "runtime.o", (char *) NULL);
 
         // unreachable
         assert (false);
@@ -812,7 +974,11 @@ compile_expression (const char *e)
   }
 
   // remove input file and close port
-  unlink (itemplate);
+    if (save_temps_p)
+    printf ("Temporary asm source kept at `%s'\n", asmtmp);
+  else
+    unlink (asmtmp);
+  free (asmtmp);
 
   // We have compiled the linked library so we are ready to
   // dynamically load the library
@@ -840,6 +1006,9 @@ compile_expression (const char *e)
   }
 
   // delete output file
-  unlink (otemplate);
+  if (save_temps_p)
+    printf ("Temporary shared object kept at `%s'\n", otemplate);
+  else
+    unlink (otemplate);
 }
